Move to file_system.h
